move out featureExtractors and subjectChangers in own api files
moreover:
- resolve TODO regarding `to + infinitive` schema in api-fluent-jvm
- improve all samples in api-fluent regarding
  fail-fast vs. non-fail-fast and formatting
