+++ apis/fluent-en_GB/atrium-api-fluent-en_GB-common/src/main/kotlin/ch/tutteli/atrium/api/fluent/en_GB/arraySubjectChangers.kt
@@ -13,7 +13,7 @@ import kotlin.jvm.JvmName
  *
  * @return The newly created [Expect] for the transformed subject.
  *
- * @sample ch.tutteli.atrium.api.fluent.en_GB.samples.ArrayExpectationSamples.asListFeature
+ * @sample ch.tutteli.atrium.api.fluent.en_GB.samples.ArraySubjectChangerSamples.asListFeature
  *
  * @since 0.9.0
  */
@@ -29,7 +29,7 @@ fun <E> Expect<out Array<out E>>.asList(): Expect<List<E>> =
  *
  * @return The newly created [Expect] for the transformed subject.
  *
- * @sample ch.tutteli.atrium.api.fluent.en_GB.samples.ArrayExpectationSamples.asList
+ * @sample ch.tutteli.atrium.api.fluent.en_GB.samples.ArraySubjectChangerSamples.asList
  *
  * @since 0.9.0
  */
@@ -45,7 +45,7 @@ fun <E> Expect<Array<E>>.asList(assertionCreator: Expect<List<E>>.() -> Unit): E
  *
  * @return The newly created [Expect] for the transformed subject.
  *
- * @sample ch.tutteli.atrium.api.fluent.en_GB.samples.ArrayExpectationSamples.asListEOut
+ * @sample ch.tutteli.atrium.api.fluent.en_GB.samples.ArraySubjectChangerSamples.asListEOut
  *
  * @since 0.9.0
  */
@@ -61,7 +61,7 @@ fun <E> Expect<Array<out E>>.asList(assertionCreator: Expect<List<E>>.() -> Unit
  *
  * @return The newly created [Expect] for the transformed subject.
  *
- * @sample ch.tutteli.atrium.api.fluent.en_GB.samples.ArrayExpectationSamples.byteArrAsListFeature
+ * @sample ch.tutteli.atrium.api.fluent.en_GB.samples.ArraySubjectChangerSamples.byteArrAsListFeature
  *
  * @since 0.9.0
  */
@@ -78,7 +78,7 @@ fun Expect<ByteArray>.asList(): Expect<List<Byte>> =
  *
  * @return The newly created [Expect] for the transformed subject.
  *
- * @sample ch.tutteli.atrium.api.fluent.en_GB.samples.ArrayExpectationSamples.byteArrAsList
+ * @sample ch.tutteli.atrium.api.fluent.en_GB.samples.ArraySubjectChangerSamples.byteArrAsList
  *
  * @since 0.9.0
  */
@@ -95,7 +95,7 @@ fun Expect<ByteArray>.asList(assertionCreator: Expect<List<Byte>>.() -> Unit): E
  *
  * @return The newly created [Expect] for the transformed subject.
  *
- * @sample ch.tutteli.atrium.api.fluent.en_GB.samples.ArrayExpectationSamples.charArrAsListFeature
+ * @sample ch.tutteli.atrium.api.fluent.en_GB.samples.ArraySubjectChangerSamples.charArrAsListFeature
  *
  * @since 0.9.0
  */
@@ -112,7 +112,7 @@ fun Expect<CharArray>.asList(): Expect<List<Char>> =
  *
  * @return The newly created [Expect] for the transformed subject.
  *
- * @sample ch.tutteli.atrium.api.fluent.en_GB.samples.ArrayExpectationSamples.charArrAsList
+ * @sample ch.tutteli.atrium.api.fluent.en_GB.samples.ArraySubjectChangerSamples.charArrAsList
  *
  * @since 0.9.0
  */
@@ -129,7 +129,7 @@ fun Expect<CharArray>.asList(assertionCreator: Expect<List<Char>>.() -> Unit): E
  *
  * @return The newly created [Expect] for the transformed subject.
  *
- * @sample ch.tutteli.atrium.api.fluent.en_GB.samples.ArrayExpectationSamples.shortArrAsListFeature
+ * @sample ch.tutteli.atrium.api.fluent.en_GB.samples.ArraySubjectChangerSamples.shortArrAsListFeature
  *
  * @since 0.9.0
  */
@@ -146,7 +146,7 @@ fun Expect<ShortArray>.asList(): Expect<List<Short>> =
  *
  * @return The newly created [Expect] for the transformed subject.
  *
- * @sample ch.tutteli.atrium.api.fluent.en_GB.samples.ArrayExpectationSamples.shortArrAsList
+ * @sample ch.tutteli.atrium.api.fluent.en_GB.samples.ArraySubjectChangerSamples.shortArrAsList
  *
  * @since 0.9.0
  */
@@ -163,7 +163,7 @@ fun Expect<ShortArray>.asList(assertionCreator: Expect<List<Short>>.() -> Unit):
  *
  * @return The newly created [Expect] for the transformed subject.
  *
- * @sample ch.tutteli.atrium.api.fluent.en_GB.samples.ArrayExpectationSamples.intArrAsListFeature
+ * @sample ch.tutteli.atrium.api.fluent.en_GB.samples.ArraySubjectChangerSamples.intArrAsListFeature
  *
  * @since 0.9.0
  */
@@ -180,7 +180,7 @@ fun Expect<IntArray>.asList(): Expect<List<Int>> =
  *
  * @return The newly created [Expect] for the transformed subject.
  *
- * @sample ch.tutteli.atrium.api.fluent.en_GB.samples.ArrayExpectationSamples.intArrAsList
+ * @sample ch.tutteli.atrium.api.fluent.en_GB.samples.ArraySubjectChangerSamples.intArrAsList
  *
  * @since 0.9.0
  */
@@ -197,7 +197,7 @@ fun Expect<IntArray>.asList(assertionCreator: Expect<List<Int>>.() -> Unit): Exp
  *
  * @return The newly created [Expect] for the transformed subject.
  *
- * @sample ch.tutteli.atrium.api.fluent.en_GB.samples.ArrayExpectationSamples.longArrAsListFeature
+ * @sample ch.tutteli.atrium.api.fluent.en_GB.samples.ArraySubjectChangerSamples.longArrAsListFeature
  *
  * @since 0.9.0
  */
@@ -214,7 +214,7 @@ fun Expect<LongArray>.asList(): Expect<List<Long>> =
  *
  * @return The newly created [Expect] for the transformed subject.
  *
- * @sample ch.tutteli.atrium.api.fluent.en_GB.samples.ArrayExpectationSamples.longArrAsList
+ * @sample ch.tutteli.atrium.api.fluent.en_GB.samples.ArraySubjectChangerSamples.longArrAsList
  *
  * @since 0.9.0
  */
@@ -231,7 +231,7 @@ fun Expect<LongArray>.asList(assertionCreator: Expect<List<Long>>.() -> Unit): E
  *
  * @return The newly created [Expect] for the transformed subject.
  *
- * @sample ch.tutteli.atrium.api.fluent.en_GB.samples.ArrayExpectationSamples.floatArrAsListFeature
+ * @sample ch.tutteli.atrium.api.fluent.en_GB.samples.ArraySubjectChangerSamples.floatArrAsListFeature
  *
  * @since 0.9.0
  */
@@ -248,7 +248,7 @@ fun Expect<FloatArray>.asList(): Expect<List<Float>> =
  *
  * @return The newly created [Expect] for the transformed subject.
  *
- * @sample ch.tutteli.atrium.api.fluent.en_GB.samples.ArrayExpectationSamples.floatArrAsList
+ * @sample ch.tutteli.atrium.api.fluent.en_GB.samples.ArraySubjectChangerSamples.floatArrAsList
  *
  * @since 0.9.0
  */
@@ -265,7 +265,7 @@ fun Expect<FloatArray>.asList(assertionCreator: Expect<List<Float>>.() -> Unit):
  *
  * @return The newly created [Expect] for the transformed subject.
  *
- * @sample ch.tutteli.atrium.api.fluent.en_GB.samples.ArrayExpectationSamples.doubleArrAsListFeature
+ * @sample ch.tutteli.atrium.api.fluent.en_GB.samples.ArraySubjectChangerSamples.doubleArrAsListFeature
  *
  * @since 0.9.0
  */
@@ -282,7 +282,7 @@ fun Expect<DoubleArray>.asList(): Expect<List<Double>> =
  *
  * @return The newly created [Expect] for the transformed subject.
  *
- * @sample ch.tutteli.atrium.api.fluent.en_GB.samples.ArrayExpectationSamples.doubleArrAsList
+ * @sample ch.tutteli.atrium.api.fluent.en_GB.samples.ArraySubjectChangerSamples.doubleArrAsList
  *
  * @since 0.9.0
  */
@@ -299,7 +299,7 @@ fun Expect<DoubleArray>.asList(assertionCreator: Expect<List<Double>>.() -> Unit
  *
  * @return The newly created [Expect] for the transformed subject.
  *
- * @sample ch.tutteli.atrium.api.fluent.en_GB.samples.ArrayExpectationSamples.booleanArrAsListFeature
+ * @sample ch.tutteli.atrium.api.fluent.en_GB.samples.ArraySubjectChangerSamples.booleanArrAsListFeature
  *
  * @since 0.9.0
  */
@@ -316,7 +316,7 @@ fun Expect<BooleanArray>.asList(): Expect<List<Boolean>> =
  *
  * @return The newly created [Expect] for the transformed subject.
  *
- * @sample ch.tutteli.atrium.api.fluent.en_GB.samples.ArrayExpectationSamples.booleanArrAsList
+ * @sample ch.tutteli.atrium.api.fluent.en_GB.samples.ArraySubjectChangerSamples.booleanArrAsList
  *
  * @since 0.9.0
  */
