+++ apis/fluent-en_GB/atrium-api-fluent-en_GB-common/src/main/kotlin/ch/tutteli/atrium/api/fluent/en_GB/collectionFeatureExtractors.kt
@@ -1,0 +1,28 @@
+package ch.tutteli.atrium.api.fluent.en_GB
+
+import ch.tutteli.atrium.creating.Expect
+import ch.tutteli.atrium.logic.*
+import ch.tutteli.kbox.identity
+
+/**
+ * Creates an [Expect] for the property [Collection.size] of the subject of `this` expectation,
+ * so that further fluent calls are assertions about it.
+ *
+ * @return The newly created [Expect] for the extracted feature.
+ *
+ * @sample ch.tutteli.atrium.api.fluent.en_GB.samples.CollectionFeatureExtractorSamples.sizeFeature
+ */
+val <T : Collection<*>> Expect<T>.size: Expect<Int>
+    get() = _logic.size(::identity).transform()
+
+/**
+ * Expects that the property [Collection.size] of the subject of `this` expectation
+ * holds all assertions the given [assertionCreator] creates for it and
+ * returns an [Expect] for the current subject of `this` expectation.
+ *
+ * @return an [Expect] for the subject of `this` expectation.
+ *
+ * @sample ch.tutteli.atrium.api.fluent.en_GB.samples.CollectionFeatureExtractorSamples.size
+ */
+fun <E, T : Collection<E>> Expect<T>.size(assertionCreator: Expect<Int>.() -> Unit): Expect<T> =
+    _logic.size(::identity).collectAndAppend(assertionCreator)
