+++ apis/fluent-en_GB/atrium-api-fluent-en_GB-common/src/main/kotlin/ch/tutteli/atrium/api/fluent/en_GB/iterableExpectations.kt
@@ -3,15 +3,11 @@ package ch.tutteli.atrium.api.fluent.en_GB
 import ch.tutteli.atrium.creating.Expect
 import ch.tutteli.atrium.logic.*
 import ch.tutteli.atrium.logic.creating.iterable.contains.IterableLikeContains
-import ch.tutteli.atrium.logic.creating.iterable.contains.creators.entriesInOrderOnly
-import ch.tutteli.atrium.logic.creating.iterable.contains.creators.valuesInOrderOnly
-import ch.tutteli.atrium.logic.creating.iterablelike.contains.reporting.InOrderOnlyReportingOptions
 import ch.tutteli.atrium.logic.creating.iterable.contains.searchbehaviours.NoOpSearchBehaviour
 import ch.tutteli.atrium.logic.creating.iterable.contains.searchbehaviours.NotSearchBehaviour
 import ch.tutteli.atrium.logic.creating.iterable.contains.steps.NotCheckerStep
+import ch.tutteli.atrium.logic.creating.iterablelike.contains.reporting.InOrderOnlyReportingOptions
 import ch.tutteli.atrium.logic.creating.typeutils.IterableLike
-import ch.tutteli.atrium.logic.utils.toVarArg
-import ch.tutteli.kbox.glue
 import ch.tutteli.kbox.identity
 
 /**
@@ -330,75 +326,3 @@ fun <E : Any, T : Iterable<E?>> Expect<T>.toHaveElementsAndAll(assertionCreatorO
  */
 fun <E, T : Iterable<E>> Expect<T>.toHaveElementsAndNoDuplicates(): Expect<T> =
     _logicAppend { containsNoDuplicates(::identity) }
-
-/**
- * Creates an [Expect] for the result of calling `min()` on the subject of `this` expectation,
- * so that further fluent calls are assertions about it.
- *
- * @return The newly created [Expect] for the extracted feature.
- *
- * @since 0.9.0
- */
-fun <E : Comparable<E>, T : Iterable<E>> Expect<T>.min(): Expect<E> =
-    _logic.min(::identity).transform()
-
-/**
- * Expects that the result of calling `min()` on the subject of `this` expectation
- * holds all assertions the given [assertionCreator] creates for it and
- * returns an [Expect] for the current subject of `this` expectation.
- *
- * @return an [Expect] for the subject of `this` expectation.
- *
- * @since 0.9.0
- */
-fun <E : Comparable<E>, T : Iterable<E>> Expect<T>.min(assertionCreator: Expect<E>.() -> Unit): Expect<T> =
-    _logic.min(::identity).collectAndAppend(assertionCreator)
-
-/**
- * Creates an [Expect] for the result of calling `max()` on the subject of `this` expectation,
- * so that further fluent calls are assertions about it.
- *
- * @return The newly created [Expect] for the extracted feature.
- *
- * @since 0.9.0
- */
-fun <E : Comparable<E>, T : Iterable<E>> Expect<T>.max(): Expect<E> =
-    _logic.max(::identity).transform()
-
-/**
- * Expects that the result of calling `max()` on  the subject of `this` expectation
- * holds all assertions the given [assertionCreator] creates for it and
- * returns an [Expect] for the current subject of `this` expectation.
- *
- * @return an [Expect] for the subject of `this` expectation.
- *
- * @since 0.9.0
- */
-fun <E : Comparable<E>, T : Iterable<E>> Expect<T>.max(assertionCreator: Expect<E>.() -> Unit): Expect<T> =
-    _logic.max(::identity).collectAndAppend(assertionCreator)
-
-/**
- * Turns `Expect<E, T : Iterable<E>>` into `Expect<List<E>`.
- *
- * The transformation as such is not reflected in reporting.
- * Use `feature { f(it::toList) }` if you want to show the transformation in reporting.
- *
- * @return The newly created [Expect] for the transformed subject.
- *
- * @since 0.14.0
- */
-fun <E, T : Iterable<E>> Expect<T>.asList(): Expect<List<E>> = _logic.changeSubject.unreported { it.toList() }
-
-/**
- * Expects that the subject of `this` expectation holds all assertions the given [assertionCreator] creates for
- * the subject as [List].
- *
- * The transformation as such is not reflected in reporting.
- * Use `feature of({ f(it::toList) }, assertionCreator)` if you want to show the transformation in reporting.
- *
- * @return an [Expect] for the subject of `this` expectation.
- *
- * @since 0.14.0
- */
-fun <E, T : Iterable<E>> Expect<T>.asList(assertionCreator: Expect<List<E>>.() -> Unit): Expect<T> =
-    apply { asList()._logic.appendAsGroup(assertionCreator) }
