+++ apis/fluent-en_GB/atrium-api-fluent-en_GB-common/src/main/kotlin/ch/tutteli/atrium/api/fluent/en_GB/iterableSubjectChangers.kt
@@ -1,0 +1,31 @@
+package ch.tutteli.atrium.api.fluent.en_GB
+
+import ch.tutteli.atrium.creating.Expect
+import ch.tutteli.atrium.logic._logic
+import ch.tutteli.atrium.logic.changeSubject
+
+/**
+ * Turns `Expect<E, T : Iterable<E>>` into `Expect<List<E>`.
+ *
+ * The transformation as such is not reflected in reporting.
+ * Use `feature { f(it::toList) }` if you want to show the transformation in reporting.
+ *
+ * @return The newly created [Expect] for the transformed subject.
+ *
+ * @since 0.14.0
+ */
+fun <E, T : Iterable<E>> Expect<T>.asList(): Expect<List<E>> = _logic.changeSubject.unreported { it.toList() }
+
+/**
+ * Expects that the subject of `this` expectation holds all assertions the given [assertionCreator] creates for
+ * the subject as [List].
+ *
+ * The transformation as such is not reflected in reporting.
+ * Use `feature of({ f(it::toList) }, assertionCreator)` if you want to show the transformation in reporting.
+ *
+ * @return an [Expect] for the subject of `this` expectation.
+ *
+ * @since 0.14.0
+ */
+fun <E, T : Iterable<E>> Expect<T>.asList(assertionCreator: Expect<List<E>>.() -> Unit): Expect<T> =
+    apply { asList()._logic.appendAsGroup(assertionCreator) }
