+++ apis/fluent-en_GB/atrium-api-fluent-en_GB-common/src/main/kotlin/ch/tutteli/atrium/api/fluent/en_GB/mapSubjectChangers.kt
@@ -1,0 +1,33 @@
+package ch.tutteli.atrium.api.fluent.en_GB
+
+import ch.tutteli.atrium.creating.Expect
+import ch.tutteli.atrium.logic._logic
+import ch.tutteli.atrium.logic.changeSubject
+
+/**
+ * Turns `Expect<Map<K, V>>` into `Expect<Set<Map.Entry<K, V>>>`.
+ *
+ * The transformation as such is not reflected in reporting.
+ * Use `feature { f(it::entries) }` if you want to show the transformation in reporting.
+ *
+ * @return The newly created [Expect] for the transformed subject.
+ *
+ * @sample ch.tutteli.atrium.api.fluent.en_GB.samples.MapSubjectChangerSamples.asEntriesFeature
+ */
+fun <K, V, T : Map<out K, V>> Expect<T>.asEntries(): Expect<Set<Map.Entry<K, V>>> =
+    _logic.changeSubject.unreported { it.entries }
+
+/**
+ * Turns `Expect<Map<K, V>>` into `Expect<Set<Map.Entry<K, V>>>` and expects that it holds all assertions the given
+ * [assertionCreator] creates for it.
+ *
+ * The transformation as such is not reflected in reporting.
+ * Use `feature { f(it::entries) }` if you want to show the transformation in reporting.
+ *
+ * @return an [Expect] for the subject of `this` expectation.
+ *
+ * @sample ch.tutteli.atrium.api.fluent.en_GB.samples.MapSubjectChangerSamples.asEntries
+ */
+fun <K, V, T : Map<out K, V>> Expect<T>.asEntries(
+    assertionCreator: Expect<Set<Map.Entry<K, V>>>.() -> Unit
+): Expect<T> = apply { asEntries()._logic.appendAsGroup(assertionCreator) }
