+++ apis/fluent-en_GB/atrium-api-fluent-en_GB-common/src/test/kotlin/ch/tutteli/atrium/api/fluent/en_GB/samples/CollectionExpectationSamples.kt
@@ -32,49 +32,4 @@ class CollectionExpectationSamples {
             expect(listOf(1, 2, 3)).toHaveSize(1)
         }
     }
-
-    @Test
-    fun sizeFeature() {
-        expect(listOf(1, 2, 3))
-            .size               // subject is now of type Int (actually 3)
-            .toBeGreaterThan(1) // subject is still of type Int (still 3)
-            .toBeLessThan(4)
-
-        fails {
-            expect(listOf(1, 2, 3))
-                .size
-                .toBeLessThan(1)    // fails
-                .toBeGreaterThan(4) // not reported because `isLessThan(1)` already fails
-            //                         use `size { ... }` if you want that all assertions are evaluated
-        }.message {
-            toContain("is less than: 1")
-            notToContain("is greater than: 4")
-        }
-    }
-
-    @Test
-    fun size() {
-        expect(listOf(1, 2, 3))
-            .size { // subject inside this block is of type Int (actually 3)
-                toBeGreaterThan(1)
-            } // subject here is back to type List<Int, String>
-            .size {
-                toBeLessThan(4)
-            }
-
-        fails {
-            // all assertions are evaluated inside an assertion group block; for more details:
-            // https://github.com/robstoll/atrium#define-single-assertions-or-assertion-groups
-
-            expect(listOf(1, 2, 3))
-                .size {
-                    toBeLessThan(1)    // fails
-                    toBeGreaterThan(4) // still evaluated even though `isLessThan(1)` already fails,
-                    //                    use `.size.` if you want a fail fast behaviour
-                }
-        }.messageToContain(
-            "is less than: 1",
-            "is greater than: 4"
-        )
-    }
 }
