+++ apis/fluent-en_GB/atrium-api-fluent-en_GB-common/src/test/kotlin/ch/tutteli/atrium/api/fluent/en_GB/samples/DocumentationUtilSamples.kt
@@ -1,0 +1,26 @@
+package ch.tutteli.atrium.api.fluent.en_GB.samples
+
+import ch.tutteli.atrium.api.fluent.en_GB.*
+import ch.tutteli.atrium.api.verbs.internal.expect
+import kotlin.test.Test
+
+class DocumentationUtilSamples {
+
+    @Test
+    fun because() {
+
+        expect("filename")
+            .because("? is not allowed in file names on Windows") {
+                notToContain("?")
+            }
+
+        data class Person(val age: Int)
+        val customers = listOf(Person(21))
+
+        expect(customers).toHaveElementsAndAll {
+            because("the legal age of maturity in Switzerland is 18") {
+                feature { f(it::age) }.toBeGreaterThanOrEqualTo(18)
+            }
+        }
+    }
+}
