+++ apis/fluent-en_GB/atrium-api-fluent-en_GB-common/src/test/kotlin/ch/tutteli/atrium/api/fluent/en_GB/samples/ListFeatureExtractorSamples.kt
@@ -4,7 +4,7 @@ import ch.tutteli.atrium.api.fluent.en_GB.*
 import ch.tutteli.atrium.api.verbs.internal.expect
 import kotlin.test.Test
 
-class ListExpectationSamples {
+class ListFeatureExtractorSamples {
 
     @Test
     fun getFeature() {
@@ -17,24 +17,19 @@ class ListExpectationSamples {
 
         fails {
             expect(list)
-                .get(3)          // fails because index 3 is out of bound
-                .toBeLessThan(0) // not reported
-            //                      use `get(index) { ... }` if you want that all expectations are evaluated
-        }.message {
-            toContain("index out of bounds")
-            notToContain("is less than: 2")
+                .get(0)              // subject is now of type Int (actually 1)
+                .toBeGreaterThan(2)  // fails
+                .toBeLessThan(0)     // not reported because `toBeGreaterThan` already fails
+            //                          use `.get(index) { ... }` if you want that all expectations are evaluated
         }
 
         fails {
             expect(list)
-                .get(0)
-                .toBeGreaterThan(2)  // fails
-                .toBeLessThan(0)     // not reported because `isGreaterThan(2)` already fails
-            //                          use `get(index) { ... }` if you want that all expectations are evaluated
-        }.message {
-            toContain("is greater than: 2")
-            notToContain("is less than: 0")
+                .get(3)          // fails because index 3 is out of bound
+                .toBeLessThan(0) // not reported because `get` already fails
+            //                      use `.get(index) { ... }` if you want that all expectations are evaluated
         }
+
     }
 
     @Test
@@ -55,15 +50,18 @@ class ListExpectationSamples {
             // all expectations are evaluated inside an expectation group block; for more details:
             // https://github.com/robstoll/atrium#define-single-assertions-or-assertion-groups
 
-            expect(list)
-                .get(0) {
-                    toBeGreaterThan(2) // fails
-                    toBeLessThan(0)    // still evaluated even though `isGreaterThan(2)` already fails,
-                    //                    use `.get(index).` if you want a fail fast behaviour
-                }
-        }.messageToContain(
-            "is greater than: 2",
-            "is less than: 0"
-        )
+            expect(list).get(0) {
+                toBeGreaterThan(2) // fails
+                toBeLessThan(0)    // still evaluated even though `toBeGreaterThan` already fails,
+                //                    use `.get(index).` if you want a fail fast behaviour
+            }
+        }
+
+
+        fails { // because index 3 is out of bound, but since we use a block...
+            expect(list).get(3) {
+                toBeLessThan(0) // ...reporting mentions that the element at index 3 was expected `to be less than: 0`
+            }
+        }
     }
 }
