+++ apis/fluent-en_GB/atrium-api-fluent-en_GB-jvm/src/main/kotlin/ch/tutteli/atrium/api/fluent/en_GB/pathAssertions.kt
@@ -89,154 +89,6 @@ fun <T : Path> Expect<T>.exists(): Expect<T> =
 fun <T : Path> Expect<T>.existsNot(): Expect<T> =
     _logicAppend { existsNot() }
 
-//TODO move to pathExpectations.kt with 0.18.0
-/**
- * Creates an [Expect] for the property [Path.extension][ch.tutteli.niok.extension]
- * (provided via [niok](https://github.com/robstoll/niok)) of the subject of `this` expectation,
- * so that further fluent calls are assertions about it.
- *
- * @return The newly created [Expect] for the extracted feature.
- *
- * @sample ch.tutteli.atrium.api.fluent.en_GB.samples.PathExpectationSamples.pathExtensionFeature
- *
- * @since 0.9.0
- */
-val <T : Path> Expect<T>.extension: Expect<String>
-    get() = _logic.extension().transform()
-
-//TODO move to pathExpectations.kt with 0.18.0
-/**
- * Expects that the property [Path.extension][ch.tutteli.niok.extension]
- * (provided via [niok](https://github.com/robstoll/niok)) of the subject of `this` expectation
- * holds all assertions the given [assertionCreator] creates for it and
- * returns an [Expect] for the current subject of `this` expectation.
- *
- * @return an [Expect] for the subject of `this` expectation.
- *
- * @sample ch.tutteli.atrium.api.fluent.en_GB.samples.PathExpectationSamples.pathExtension
- *
- * @since 0.9.0
- */
-fun <T : Path> Expect<T>.extension(assertionCreator: Expect<String>.() -> Unit): Expect<T> =
-    _logic.extension().collectAndAppend(assertionCreator)
-
-//TODO move to pathExpectations.kt with 0.18.0
-/**
- * Creates an [Expect] for the property [Path.fileNameAsString][ch.tutteli.niok.fileNameAsString]
- * (provided via [niok](https://github.com/robstoll/niok)) of the subject of `this` expectation,
- * so that further fluent calls are assertions about it.
- *
- * @return The newly created [Expect] for the extracted feature.
- *
- * @sample ch.tutteli.atrium.api.fluent.en_GB.samples.PathExpectationSamples.fileNameFeature
- *
- * @since 0.9.0
- */
-val <T : Path> Expect<T>.fileName: Expect<String>
-    get() = _logic.fileName().transform()
-
-//TODO move to pathExpectations.kt with 0.18.0
-/**
- * Expects that the property [Path.fileNameAsString][ch.tutteli.niok.fileNameAsString]
- * (provided via [niok](https://github.com/robstoll/niok)) of the subject of `this` expectation
- * holds all assertions the given [assertionCreator] creates for it and
- * returns an [Expect] for the current subject of `this` expectation.
- *
- * @return an [Expect] for the subject of `this` expectation.
- *
- * @sample ch.tutteli.atrium.api.fluent.en_GB.samples.PathExpectationSamples.fileName
- *
- * @since 0.9.0
- */
-fun <T : Path> Expect<T>.fileName(assertionCreator: Expect<String>.() -> Unit): Expect<T> =
-    _logic.fileName().collectAndAppend(assertionCreator)
-
-//TODO move to pathExpectations.kt with 0.18.0
-/**
- * Creates an [Expect] for the property [Path.fileNameWithoutExtension][ch.tutteli.niok.fileNameWithoutExtension]
- * (provided via [niok](https://github.com/robstoll/niok)) of the subject of `this` expectation,
- * so that further fluent calls are assertions about it.
- *
- * @return The newly created [Expect] for the extracted feature.
- *
- * @sample ch.tutteli.atrium.api.fluent.en_GB.samples.PathExpectationSamples.fileNameWithoutExtensionFeature
- *
- * @since 0.9.0
- */
-val <T : Path> Expect<T>.fileNameWithoutExtension: Expect<String>
-    get() = _logic.fileNameWithoutExtension().transform()
-
-//TODO move to pathExpectations.kt with 0.18.0
-/**
- * Expects that the property [Path.fileNameWithoutExtension][ch.tutteli.niok.fileNameWithoutExtension]
- * (provided via [niok](https://github.com/robstoll/niok))
- * of the subject of `this` expectation holds all assertions the given [assertionCreator] creates for it
- * and returns an [Expect] for the current subject of `this` expectation.
- *
- * @return an [Expect] for the subject of `this` expectation.
- *
- * @sample ch.tutteli.atrium.api.fluent.en_GB.samples.PathExpectationSamples.fileNameWithoutExtension
- *
- * @since 0.9.0
- */
-fun <T : Path> Expect<T>.fileNameWithoutExtension(assertionCreator: Expect<String>.() -> Unit): Expect<T> =
-    _logic.fileNameWithoutExtension().collectAndAppend(assertionCreator)
-
-//TODO move to pathExpectations.kt with 0.18.0
-/**
- * Expects that this [Path] has a [parent][Path.getParent] and creates an [Expect] for it,
- * so that further fluent calls are assertions about it.
- *
- * @return The newly created [Expect] for the extracted feature.
- *
- * @sample ch.tutteli.atrium.api.fluent.en_GB.samples.PathExpectationSamples.parentFeature
- *
- * @since 0.9.0
- */
-val <T : Path> Expect<T>.parent: Expect<Path>
-    get() = _logic.parent().transform()
-
-//TODO move to pathExpectations.kt with 0.18.0
-/**
- * Expects that this [Path] has a [parent][Path.getParent], that the parent holds all assertions the
- * given [assertionCreator] creates for it and returns an [Expect] for the current subject of `this` expectation.
- *
- * @return an [Expect] for the subject of `this` expectation.
- *
- * @sample ch.tutteli.atrium.api.fluent.en_GB.samples.PathExpectationSamples.parent
- *
- * @since 0.9.0
- */
-fun <T : Path> Expect<T>.parent(assertionCreator: Expect<Path>.() -> Unit): Expect<T> =
-    _logic.parent().collectAndAppend(assertionCreator)
-
-//TODO move to pathExpectations.kt with 0.18.0
-/**
- * Expects that [other] resolves against this [Path] and creates an [Expect] for the resolved [Path]
- * so that further fluent calls are assertions about it.
- *
- * @return The newly created [Expect] for the extracted feature.
- *
- * @sample ch.tutteli.atrium.api.fluent.en_GB.samples.PathExpectationSamples.resolveFeature
- *
- * @since 0.10.0
- */
-fun <T : Path> Expect<T>.resolve(other: String): Expect<Path> =
-    _logic.resolve(other).transform()
-
-//TODO move to pathExpectations.kt with 0.18.0
-/**
- * Expects that [other] resolves against this [Path], that the resolved [Path] holds all assertions the
- * given [assertionCreator] creates for it and returns an [Expect] for the current subject of `this` expectation.
- *
- * @return an [Expect] for the subject of `this` expectation.
- *
- * @sample ch.tutteli.atrium.api.fluent.en_GB.samples.PathExpectationSamples.resolve
- *
- * @since 0.10.0
- */
-fun <T : Path> Expect<T>.resolve(other: String, assertionCreator: Expect<Path>.() -> Unit): Expect<T> =
-    _logic.resolve(other).collectAndAppend(assertionCreator)
 
 /**
  * Expects that the subject of `this` expectation (a [Path]) is readable;
