--- apis/fluent-en_GB/atrium-api-fluent-en_GB-jvm/src/test/kotlin/ch/tutteli/atrium/api/fluent/en_GB/FloatingPointWithErrorToleranceExpectationsJvmSpec.kt
@@ -1,27 +1,0 @@
-//TODO remove file with 0.18.0
-@file:Suppress("DEPRECATION")
-package ch.tutteli.atrium.api.fluent.en_GB
-
-import ch.tutteli.atrium.creating.Expect
-import java.math.BigDecimal
-import kotlin.reflect.KFunction3
-
-class FloatingPointWithErrorToleranceExpectationsJvmSpec :
-    ch.tutteli.atrium.specs.integration.FloatingPointWithErrorToleranceExpectationsJvmSpec(
-        toBeWithErrorToleranceBigDecimalPair()
-    ) {
-    companion object {
-
-        private val toBeWithErrorToleranceBigDecimalFun: KFunction3<Expect<BigDecimal>, BigDecimal, BigDecimal, Expect<BigDecimal>> =
-            Expect<BigDecimal>::toBeWithErrorTolerance
-
-        fun toBeWithErrorToleranceBigDecimalPair() =
-            "${toBeWithErrorToleranceBigDecimalFun.name} for BigDecimal" to Companion::toBeWithErrorToleranceBigDecimal
-
-        private fun toBeWithErrorToleranceBigDecimal(
-            expect: Expect<BigDecimal>,
-            expected: BigDecimal,
-            tolerance: BigDecimal
-        ) = expect.toBeWithErrorTolerance(expected, tolerance)
-    }
-}
